roia-suite/backend/requirements.txt
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
python-dotenv>=0.19.0
openai>=1.0.0
pydantic>=1.8.0
httpx>=0.23.0
python-multipart>=0.0.5
requests>=2.25.0
openpyxl>=3.0.0
pandas>=1.3.0
jira>=3.5.0
